add config and endpoints to Deposit, update classes to use embedded object of User and House id (UserHouse)

+++ src/main/java/eeet2582/realestatemgt/appuser/UserService.java
@@ -26,6 +26,9 @@ public class UserService {
     }
 
     public void deleteById(Long userId) {
+        if (!userRepository.existsById(userId))
+            throw new IllegalStateException("User with userId=" + userId + " does not exist!");
+
         userRepository.deleteById(userId);
     }
 }
